pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Matheus Mendez <matheusmendez88@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.3"
mysql-connector-python = "^8.0.32"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
